# openapi.yaml
openapi: 3.0.3
info:
  title: Placement Hub API
  version: "1.0"
  description: |
    Student, recruiter and placement-cell API. Auth is a short-lived Bearer access token; the refresh token
    lives in an httpOnly cookie (`ph_refresh`, path `/auth`). Errors always look like
    `{"error": {"code": "...", "message": "...", "fields": {"field": "reason"}}}`.

    A student who accepts an offer is PLACED: applying to jobs or registering for drives then fails with
    `403 already_placed`, while their history stays readable.
servers:
  - url: http://localhost:8080
security:
  - bearer: []
...
paths:
...
  /jobs:
    get:
      tags: [Jobs]
      summary: 'Open jobs with per-student eligibility (`eligible`, `applied`, `canApply`, `blockReason`)'
      parameters:
        - {name: type, in: query, schema: {type: string, enum: [All, Full-time, Internship]}}
        - {name: q, in: query, schema: {type: string}}
        - {name: eligible, in: query, schema: {type: boolean}}
        - {name: sort, in: query, schema: {type: string, enum: [deadline, ctc]}}
        - {$ref: "#/components/parameters/page"}
        - {$ref: "#/components/parameters/limit"}
      responses: {"200": {description: Page of jobs, content: {application/json: {schema: {$ref: "#/components/schemas/JobPage"}}}}}
  /jobs/{id}:
    get:
      tags: [Jobs]
      summary: 'One job (closed jobs stay visible to students who applied)'
      parameters: [{$ref: "#/components/parameters/id"}]
      responses: {"200": {description: Job, content: {application/json: {schema: {$ref: "#/components/schemas/StudentJob"}}}}, "404": {$ref: "#/components/responses/Error"}}
...
  /company/jobs:
    get: {tags: [Company], summary: "All own postings in every status", responses: {"200": {description: Items}}}
    post:
      tags: [Company]
      summary: 'Create a draft; `"submit": true` also sends it for approval'
      requestBody: {required: true, content: {application/json: {schema: {$ref: "#/components/schemas/JobInput"}}}}
      responses: {"201": {description: Job}, "422": {$ref: "#/components/responses/Error"}}
  /company/jobs/{id}:
    get: {tags: [Company], parameters: [{$ref: "#/components/parameters/id"}], summary: "One own posting", responses: {"200": {description: Job}}}
    put: {tags: [Company], parameters: [{$ref: "#/components/parameters/id"}], summary: "Edit. Draft/Rejected: everything. Open: only deadline, openings, skills, description (409 criteria_locked otherwise).", responses: {"200": {description: Job}}}
    delete: {tags: [Company], parameters: [{$ref: "#/components/parameters/id"}], summary: "Delete a draft", responses: {"204": {description: Deleted}}}
  /company/jobs/{id}/submit:
    post: {tags: [Company], parameters: [{$ref: "#/components/parameters/id"}], summary: "Send for approval (422 company_not_approved until the company is approved)", responses: {"200": {description: Job}}}
  /company/jobs/{id}/close:
    post: {tags: [Company], parameters: [{$ref: "#/components/parameters/id"}], summary: "Stop taking applications", responses: {"200": {description: Job}}}
...
  /admin/students:
    get:
      tags: [Admin]
      summary: 'Students with derived status (Placed / In process / Unplaced)'
      parameters:
        - {name: q, in: query, schema: {type: string}}
        - {name: branch, in: query, schema: {type: string}}
        - {name: status, in: query, schema: {type: string, enum: [All, Placed, In process, Unplaced]}}
        - {$ref: "#/components/parameters/page"}
        - {$ref: "#/components/parameters/limit"}
      responses: {"200": {description: Page}}
    post: {tags: [Admin], summary: "Create a student; returns a one-time temporary password", responses: {"201": {description: Created}}}
...
  /admin/students/{id}:
    get: {tags: [Admin], parameters: [{$ref: "#/components/parameters/id"}], summary: "Student profile", responses: {"200": {description: Profile}}}
    patch: {tags: [Admin], parameters: [{$ref: "#/components/parameters/id"}], summary: "Edit academics (name, branch, year, cgpa, backlogs)", responses: {"200": {description: Profile}}}
...
  /admin/students/{id}/activate:
    post: {tags: [Admin], parameters: [{$ref: "#/components/parameters/id"}], summary: "Enable login", responses: {"204": {description: Done}}}
  /admin/students/{id}/deactivate:
    post: {tags: [Admin], parameters: [{$ref: "#/components/parameters/id"}], summary: "Disable login and sign out everywhere", responses: {"204": {description: Done}}}
  /admin/students/{id}/reset-password:
    post: {tags: [Admin], parameters: [{$ref: "#/components/parameters/id"}], summary: "Issue a temporary password", responses: {"200": {description: "{tempPassword}"}}}
...
  /admin/jobs:
    get: {tags: [Admin], parameters: [{name: status, in: query, schema: {type: string, enum: [All, Pending, Open, Closed, Rejected]}}], summary: "Submitted jobs (drafts are private)", responses: {"200": {description: Items}}}
  /admin/jobs/{id}/approve:
    post: {tags: [Admin], parameters: [{$ref: "#/components/parameters/id"}], summary: "Make a pending job live and notify exactly the eligible students ({job, studentsNotified})", responses: {"200": {description: Result}}}
  /admin/jobs/{id}/reject:
    post: {tags: [Admin], parameters: [{$ref: "#/components/parameters/id"}], summary: "Send back with a required {reason}", responses: {"200": {description: Job}}}
  /admin/jobs/{id}/close:
    post: {tags: [Admin], parameters: [{$ref: "#/components/parameters/id"}], summary: "Close any live job", responses: {"200": {description: Job}}}
...
components:
  securitySchemes:
    bearer: {type: http, scheme: bearer, bearerFormat: JWT}
  parameters:
    id: {name: id, in: path, required: true, schema: {type: string, format: uuid}}
    page: {name: page, in: query, schema: {type: integer, minimum: 1, default: 1}}
    limit: {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 100, default: 20}}
  responses:
    Error:
      description: Error
      content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: {type: string, example: already_placed}
            message: {type: string}
            fields: {type: object, additionalProperties: {type: string}}
...
    Job:
      type: object
      properties:
        id: {type: string}
        companyId: {type: string}
        company: {type: string}
        color: {type: string}
        role: {type: string}
        type: {type: string, enum: [Full-time, Internship]}
        location: {type: string}
        ctc: {type: number}
        minCgpa: {type: number}
        branches: {type: array, items: {type: string}}
        skills: {type: array, items: {type: string}}
        deadline: {type: string, format: date}
        openings: {type: integer}
        applicants: {type: integer}
        status: {type: string, enum: [Draft, Pending, Open, Closed, Rejected]}
        description: {type: string}
        allowBacklogs: {type: boolean}
        rejectReason: {type: string}
    StudentJob:
      allOf:
        - $ref: "#/components/schemas/Job"
        - type: object
          properties:
            eligible: {type: boolean}
            applied: {type: boolean}
            applicationId: {type: string, nullable: true}
            canApply: {type: boolean}
            blockReason: {type: object, nullable: true, properties: {code: {type: string}, message: {type: string}}}
    JobPage:
      type: object
      properties: {items: {type: array, items: {$ref: "#/components/schemas/StudentJob"}}, total: {type: integer}, page: {type: integer}, limit: {type: integer}}
    JobInput:
      type: object
      required: [role, type, location, ctc, branches, deadline, openings, description]
      properties:
        role: {type: string}
        type: {type: string, enum: [Full-time, Internship]}
        location: {type: string}
        ctc: {type: number}
        minCgpa: {type: number}
        branches: {type: array, items: {type: string}}
        skills: {type: array, items: {type: string}}
        deadline: {type: string, format: date}
        openings: {type: integer}
        description: {type: string}
        allowBacklogs: {type: boolean}
        submit: {type: boolean}
...